+++ backend/src/main/java/com/vendaspedidos/entities/Categoria.java
@@ -18,7 +18,7 @@ public class Categoria implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String nome;
 	
 	@ManyToMany(mappedBy = "categorias")
@@ -27,16 +27,16 @@ public class Categoria implements Serializable{
 	public Categoria() {
 	}
 
-	public Categoria(Long id, String nome) {
+	public Categoria(Integer id, String nome) {
 		this.id = id;
 		this.nome = nome;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
